src/backend/utils/python/requirements.txt
adafruit-io==3.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
paho-mqtt==2.1.0
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
